requirements.txt
praw
requests
pymongo
pandas
python-dateutil
vaderSentiment
python-dotenv
fastapi
uvicorn[standard]
pymongo
python-dotenv
pandas
python-dateutil
